Update README.md
Part 2: Introduction to AI and Prompt Engineering

Define prompt engineering and discuss its importance in interacting with AI models.

Prompt Engineering can be defined as an engineer who craft questions or statements to get the best possible response from AI models

Provide an example of a vague prompt and then improve it by making it clear, specific, and concise. Explain why the improved prompt is more effective.

Example: LETTER OF INTENT TO RUN FOR CHAIRMAN OF OJOKORO LCDA IN 2025

Because the more specification the input is, the more relevant the result I get.